Move JobScheduler to scheduler package
